backend updated

+++ backend/controllers/productController.js
@@ -37,7 +37,7 @@ export const createProduct = async (req, res) => {
         });
     } catch (error) {
       console.log(error.message);
-      res.status(500).json({ success: false, message: "Error adding product" });
+      res.status(500).json({ success: false, message: "Server error" });
     }
 }
 
@@ -46,13 +46,17 @@ export const updateProduct = async (req, res) => {
     const id = req.params.id;
     const product = req.body;
 
+    if (mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(404).json({ success: false, message: "Invalid ID" });
+    }
+
     try {
       await Product.findByIdAndUpdate(id, product, { new: true });
       res
         .status(200)
         .json({ success: true, message: "Product updated successfully" });
     } catch (error) {
-      res.status(500).json({ success: false, message: "Product not found" });
+      res.status(500).json({ success: false, message: "Server error" });
     }
 }
 
@@ -60,12 +64,16 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const id = req.params.id;
 
+    if (mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(404).json({ success: false, message: "Invalid ID" });
+    }
+
     try {
       await Product.findByIdAndDelete(id);
       res
         .status(200)
         .json({ success: true, message: "Product deleted successfully" });
     } catch (error) {
-      res.status(500).json({ success: false, message: "Product not found" });
+      res.status(500).json({ success: false, message: "Server error" });
     }
 }
